integration: merge batch tests into a single file
As noted during the review of this PR, it's better to have all
batch-related tests in a single file for better organization
and maintainability.
