--- scylla/tests/integration/statements/batch/mod.rs
@@ -1,4 +1,0 @@
-#[expect(clippy::module_inception)]
-mod batch;
-mod large_batch_statements;
-mod silent_prepare_batch;
